Refactor tests and documentation, create swap_info fixture
