+++ README.md
@@ -171,17 +171,21 @@ $ remme atomic-swap get-info \
       --id=033402fe1346742486b15a3a9966eb5249271025fc7fb0b37ed3fdb4bcce6808 \
       --node-url=node-genesis-testnet.remme.io
 {
-    'sender_address': '112007be95c8bb240396446ec359d0d7f04d257b72aeb4ab1ecfe50cf36e400a96ab9c',
-    'receiver_address': '112007484def48e1c6b77cf784aeabcac51222e48ae14f3821697f4040247ba01558b1',
-    'amount': '10.0000',
-    'swap_id': '033402fe1346742486b15a3a9966eb5249271025fc7fb0b37ed3fdb4bcce6808',
-    'secret_lock': '0728356568862f9da0825aa45ae9d3642d64a6a732ad70b8857b2823dbf2a0b8',
-    'created_at': 1555943451,
-    'sender_address_non_local': '0xe6ca0e7c974f06471759e9a05d18b538c5ced11e',
-    'state': 'OPENED',
-    'email_address_encrypted_optional': '',
-    'secret_key': '',
-    'is_initiator': False,
+    "result": {
+        "information": {
+            "amount": "10.0000",
+            "created_at": 1556803765,
+            "email_address_encrypted_optional": "",
+            "is_initiator": false,
+            "receiver_address": "112007484def48e1c6b77cf784aeabcac51222e48ae14f3821697f4040247ba01558b1",
+            "secret_key": "",
+            "secret_lock": "0728356568862f9da0825aa45ae9d3642d64a6a732ad70b8857b2823dbf2a0b8",
+            "sender_address": "1120076ecf036e857f42129b58303bcf1e03723764a1702cbe98529802aad8514ee3cf",
+            "sender_address_non_local": "0xe6ca0e7c974f06471759e9a05d18b538c5ced11e",
+            "state": "OPENED",
+            "swap_id": "033402fe1346742486b15a3a9966eb5249271025fc7fb0b37ed3fdb4bcce6808"
+        }
+    }
 }
 ```
